+++ Lib/PvpXray/AssemblyInfo.cs
@@ -4,3 +4,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.PackageValidationSuite.Editor")]
 [assembly: InternalsVisibleTo("Unity.PackageValidationSuite.EditorTests")]
 [assembly: InternalsVisibleTo("UpmPvp")]
+[assembly: InternalsVisibleTo("UpmPvp.Tests")]