+++ Lib/PvpXray/XrayUtils.cs
@@ -1,9 +1,16 @@
 using System.IO;
+using System.Text;
 
 namespace PvpXray
 {
     static class XrayUtils
     {
+        /// Version of Encoding.UTF8 that throws on encoding errors instead of
+        /// letting them pass silently with a replacement character.
+        /// This affects conversions TO string only (Encoding.GetString), not
+        /// vice versa (GetBytes etc.), where Encoding.UTF8 remains acceptable.
+        public static readonly UTF8Encoding Utf8Strict = new UTF8Encoding(encoderShouldEmitUTF8Identifier: false, throwOnInvalidBytes: true);
+
         /// Reading streams into a byte array is limited by the maximum byte[]
         /// length. For Mono (and thus, Unity) the limit is Int32.MaxValue
         /// (2147483647) bytes (tested on 2019.2.21f1 and 2023.2.0a5), but .NET
@@ -23,6 +30,15 @@ namespace PvpXray
                 var nRead = stream.Read(buffer, offset, buffer.Length - offset);
                 if (nRead < 1) throw new IOException($"Unexpected EOF at offset {offset} of {buffer.Length}");
                 offset += nRead;
+            }
+        }
+
+        /// Read stream as a UTF-8 string (in strict mode), and dispose stream.
+        public static string ReadToString(Stream stream)
+        {
+            using (var sr = new StreamReader(stream, Utf8Strict, detectEncodingFromByteOrderMarks: false))
+            {
+                return sr.ReadToEnd();
             }
         }